wizard: rename index.ts to slice.ts with barrel re-export
